examples/edge_deployment/deploy_edge_agent.py
import os
import yaml
import subprocess
import logging

# Configure logging
logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
logger = logging.getLogger("EdgeDeployment")

# Load configuration from YAML file
CONFIG_FILE_PATH = "configure_edge.yaml"

def load_configuration(config_path):
    """
    Loads the configuration from the provided YAML file.
    """
    try:
        with open(config_path, 'r') as file:
            config = yaml.safe_load(file)
            logger.info("Configuration loaded successfully from %s", config_path)
            return config
    except Exception as e:
        logger.error("Failed to load configuration: %s", e)
        raise

# Setup network
def setup_network(config):
    """
    Configures the network based on the provided configuration.
    """
    if config['network']['use_wifi']:
        ssid = config['network']['wifi_ssid']
        password = config['network']['wifi_password']
        logger.info("Configuring WiFi network...")
        # This is a simple example and should be tailored for the specific device and OS
        command = f"nmcli dev wifi connect '{ssid}' password '{password}'"
        try:
            subprocess.run(command, shell=True, check=True)
            logger.info("WiFi network configured successfully.")
        except subprocess.CalledProcessError as e:
            logger.error("Failed to configure WiFi: %s", e)
            raise

# Setup logging
def setup_logging(config):
    """
    Sets up logging to the specified log file.
    """
    log_file_path = config['logging']['log_file_path']
    file_handler = logging.FileHandler(log_file_path)
    file_handler.setFormatter(logging.Formatter('%(asctime)s - %(levelname)s - %(message)s'))
    logger.addHandler(file_handler)
    logger.info("Logging configured. Writing logs to %s", log_file_path)

# Deploy the model
def deploy_model(config):
    """
    Deploys the AI model based on the provided configuration.
    """
    model_path = config['model']['model_storage_path']
    if not os.path.exists(model_path):
        logger.error("Model file not found at %s", model_path)
        raise FileNotFoundError(f"Model file not found at {model_path}")
    logger.info("Model loaded successfully from %s", model_path)
    # Simulating model deployment; replace with actual model loading/deployment

# Secure communication setup
def setup_security(config):
    """
    Sets up secure communication based on the provided configuration.
    """
    if config['security']['encryption']['enabled']:
        key_path = config['security']['encryption']['key_path']
        if not os.path.exists(key_path):
            logger.error("Encryption key file not found at %s", key_path)
            raise FileNotFoundError(f"Encryption key file not found at {key_path}")
        logger.info("Encryption key loaded from %s", key_path)
        # Additional encryption setup logic here

# Main deployment function
def deploy_edge_agent():
    """
    Main function to deploy the edge agent based on configuration.
    """
    try:
        config = load_configuration(CONFIG_FILE_PATH)
        setup_logging(config)
        setup_network(config)
        setup_security(config)
        deploy_model(config)
        logger.info("Edge agent deployed successfully.")
    except Exception as e:
        logger.error("Edge agent deployment failed: %s", e)
        raise

if __name__ == "__main__":
    deploy_edge_agent()